move mergeStates test to api package
